add some new functions for new ffmpeg version, test example on `ffmpeg version n4.3.2-160-gfbb9368226 Copyright (c) 2000-2021 the FFmpeg developers` ok.

+++ avutil/memory.go
@@ -15,7 +15,7 @@ import (
 )
 
 //Allocate a block of size bytes with alignment suitable for all memory accesses (including vectors if available on the CPU).
-func AvMalloc(s uint) unsafe.Pointer {
+func AvMalloc(s uintptr) unsafe.Pointer {
 	return unsafe.Pointer(C.av_malloc(C.size_t(s)))
 }
